+++ R/Performer.R
@@ -52,15 +52,15 @@ computeUnderperformer = function(data,
                                  thresholdValue = 0,
                                  thresholdTime = 90,
                                  use_latest = FALSE) {
-
-
+  
+  
   if(use_latest == TRUE & !is.numeric(thresholdValue)){
     data<-expandData(data = data,
                      expand = group,
                      expandTo = "all",
                      valueColumns = c(value,thresholdValue), latest_values = T,timestamp = timestamp, timestampFormat = timestampFormat, keepData = F)
   }
-
+  
   if(use_latest == TRUE & is.numeric(thresholdValue)){
     data<-expandData(data = data,
                      expand = group,
@@ -71,22 +71,11 @@ computeUnderperformer = function(data,
                      timestampFormat = timestampFormat,
                      keepData = F)
   }
-
+  
   ### Bring timestamp and currentTime to date format
-  if (timestampFormat == "day") {
-    data[[timestamp]] = as.Date(as.character(data[[timestamp]]))
-  }
-  else if (timestampFormat == "week") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-1")
-    data[[timestamp]] = ISOweek2date(data[[timestamp]])
-    currentTime = paste0(currentTime, "-1")
-    currentTime<-ISOweek2date(currentTime)
-  }
-  else if (timestampFormat == "month") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01")
-    currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
+  if (timestampFormat %in% c("day", "week", "month", "year")) {
+    data[[timestamp]] = as.Date(parse_iso_8601(data[[timestamp]]))
+    currentTime = as.Date(parse_iso_8601(currentTime))
   }
   else if (timestampFormat == "quarter") {
     data[[timestamp]] = paste0(substr(data[[timestamp]],1,4), "-", as.numeric(substr(data[[timestamp]],7,7))*3, "-01")
@@ -94,17 +83,10 @@ computeUnderperformer = function(data,
     currentTime = paste0(substr(currentTime,1,4), "-", as.numeric(substr(currentTime,7,7))*3, "-01")
     currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
   }
-  else if (timestampFormat == "year") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01-01")
-    currentTime = as.Date(currentTime, format = "%Y-%m-%d")
-  }
-
-
+  
   ### Select only Data in the past of currentTime
   data = data[data[[timestamp]] <= currentTime, ]
-
+  
   ### Determine last timestamp when thresholdValue was overtaken
   if (!is.numeric(thresholdValue)) {
     currentdata = data[which(data[, timestamp] == currentTime), c(group, value, timestamp, thresholdValue)]
@@ -127,61 +109,66 @@ computeUnderperformer = function(data,
     ),
     by = eval(get("group"))]
   }
-
+  
   result = left_join(currentdata, lastexceed, by = group)
   result = result[order(result$lastover, -result$moves), ]
-
+  
   ### Subset data for condition thresholdTime and determine timespan in aggregation of timestampFormat
   if(timestampFormat == "day"){
     result = result[result$lastover <= as.Date(currentTime) - thresholdTime, ]
     if(nrow(result)>0){
-      result$toolowindays = result[[timestamp]] - result$lastover
+      result$toolowindays = paste(result[[timestamp]] - result$lastover, "days")
     }
   }
   else if(timestampFormat =="week"){
-    result = result[result$lastover <= as.Date(currentTime) - as.difftime(thresholdTime, units="weeks"), ]
+    result = result[result$lastover <= as.Date(currentTime) %m-% weeks(thresholdTime), ]
     if(nrow(result)>0){
-      result$toolowinweeks = paste((result[[timestamp]] - result$lastover)/7,"weeks")
-      result$lastover = ISOweek(result$lastover)
-      result[[timestamp]] = ISOweek(result[[timestamp]])
+      result$toolowinweeks = paste(round((result[[timestamp]] - result$lastover)/7, digits = 2),"weeks")
+      result$lastover = paste(isoyear(result$lastover), sprintf("%02d", isoweek(result$lastover)), sep = "-W")
+      result$mindate = paste(isoyear(result$mindate), sprintf("%02d", isoweek(result$mindate)), sep = "-W")
+      result[[timestamp]] = paste(isoyear(result[[timestamp]]), sprintf("%02d", isoweek(result[[timestamp]])), sep = "-W")
     }
   }
   else if(timestampFormat == "month"){
-    result = result[result$lastover <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," months"))[2], ]
+    result = result[result$lastover <= as.Date(currentTime) %m-% months(thresholdTime), ]
     if(nrow(result)>0){
-      result$toolowinmonth = paste((year(result[[timestamp]]) - year(result$lastover))*12 + (month(result[[timestamp]]) - month(result$lastover)),"month")
-      result$lastover = format(result$lastover, "%Y-%m")
-      result[[timestamp]] = format(result[[timestamp]], "%Y-%m")
+      result$toolowinmonth = paste(interval(result$lastover, result[[timestamp]]) %/% months(1), "months")
+      result$lastover = paste(year(result$lastover), sprintf("%02d", month(result$lastover)), sep = "-")
+      result$mindate = paste(year(result$mindate), sprintf("%02d", month(result$mindate)), sep = "-")
+      result[[timestamp]] = paste(year(result[[timestamp]]), sprintf("%02d", month(result[[timestamp]])), sep = "-")
     }
   }
   else if(timestampFormat == "quarter"){
-    result = result[result$lastover <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," quarters"))[2], ]
+    result = result[result$lastover <= as.Date(currentTime) %m-% months(thresholdTime * 3), ]
     if(nrow(result)>0){
-      result$toolowinquarters = paste((year(result[[timestamp]]) - year(result$lastover))*4 + (quarter(result[[timestamp]]) - quarter(result$lastover)),"quaters")
+      result$toolowinquarters = paste(round((interval(result$lastover, result[[timestamp]]) %/% months(1)) /3, digits = 2 ), "quarters")
       result$lastover = paste0(year(result$lastover), "-Q", quarter(result$lastover))
+      result$mindate = paste0(year(result$mindate), "-Q", quarter(result$mindate))
       result[[timestamp]] = paste0(year(result[[timestamp]]), "-Q", quarter(result[[timestamp]]))
     }
   }
   else if(timestampFormat == "year"){
-    result = result[result$lastover <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," years"))[2], ]
+    result = result[result$lastover <= as.Date(currentTime) %m-% years(thresholdTime), ]
     if(nrow(result)>0){
       result$toolowinyears = paste(year(result[[timestamp]]) - year(result$lastover),"years")
       result$lastover = year(result$lastover)
+      result$mindate = year(result$mindate)
       result[[timestamp]] = year(result[[timestamp]])
     }
   }
-
-  if(any(result$lastover == result$mindate )){
-
+  
+  if(any(result$lastover == result$mindate)){
+    
     equals <- which(result$lastover == result$mindate)
-
+    
+    result <- data.frame(result)
     result$lastover<- as.character(result$lastover)
     result$lastover[equals] <- paste(">=", result$lastover[equals])
     result[, ncol(result)]<- as.character( result[, ncol(result)])
     result[equals, ncol(result)] <- paste(">=", result[equals, ncol(result)])
   }
   result$mindate <- NULL
-
+  
   return(result)
 }
 
@@ -239,7 +226,7 @@ computeOverperformer = function(data,
                      expandTo = "all",
                      valueColumns = c(value,thresholdValue), latest_values = T,timestamp = timestamp, timestampFormat = timestampFormat, keepData = F)
   }
-
+  
   if(use_latest == TRUE & is.numeric(thresholdValue)){
     data<-expandData(data = data,
                      expand = group,
@@ -250,23 +237,11 @@ computeOverperformer = function(data,
                      timestampFormat = timestampFormat,
                      keepData = F)
   }
-
-
+  
   ### Bring timestamp and currentTime to date format
-  if (timestampFormat == "day") {
-    data[[timestamp]] = as.Date(as.character(data[[timestamp]]))
-  }
-  else if (timestampFormat == "week") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-1")
-    data[[timestamp]] = ISOweek2date(data[[timestamp]])
-    currentTime = paste0(currentTime, "-1")
-    currentTime<-ISOweek2date(currentTime)
-  }
-  else if (timestampFormat == "month") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01")
-    currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
+  if (timestampFormat %in% c("day", "week", "month", "year")) {
+    data[[timestamp]] = as.Date(parse_iso_8601(data[[timestamp]]))
+    currentTime = as.Date(parse_iso_8601(currentTime))
   }
   else if (timestampFormat == "quarter") {
     data[[timestamp]] = paste0(substr(data[[timestamp]],1,4), "-", as.numeric(substr(data[[timestamp]],7,7))*3, "-01")
@@ -274,13 +249,7 @@ computeOverperformer = function(data,
     currentTime = paste0(substr(currentTime,1,4), "-", as.numeric(substr(currentTime,7,7))*3, "-01")
     currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
   }
-  else if (timestampFormat == "year") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01-01")
-    currentTime = as.Date(currentTime, format = "%Y-%m-%d")
-  }
-
+  
   ### Select only Data in the past of currentTime
   data = data[data[[timestamp]] <= currentTime, ]
 
@@ -314,38 +283,42 @@ computeOverperformer = function(data,
   if(timestampFormat == "day"){
     result = result[result$lastunder <= as.Date(currentTime) - thresholdTime, ]
     if(nrow(result)>0){
-      result$toohighindays = result[[timestamp]] - result$lastunder
+      result$toohighindays = paste(result[[timestamp]] - result$lastunder, "days")
     }
   }
   else if(timestampFormat =="week"){
-    result = result[result$lastunder <= as.Date(currentTime) - as.difftime(thresholdTime, units="weeks"), ]
+    result = result[result$lastunder <= as.Date(currentTime) %m-% weeks(thresholdTime), ]
     if(nrow(result)>0){
-      result$toohighinweeks = paste((result[[timestamp]] - result$lastunder)/7,"weeks")
-      result$lastunder = ISOweek(result$lastunder)
-      result[[timestamp]] = ISOweek(result[[timestamp]])
+      result$toohighinweeks = paste(round((result[[timestamp]] - result$lastunder)/7, digits = 2),"weeks")
+      result$lastunder = paste(isoyear(result$lastunder), sprintf("%02d", isoweek(result$lastunder)), sep = "-W")
+      result$mindate = paste(isoyear(result$mindate), sprintf("%02d", isoweek(result$mindate)), sep = "-W")
+      result[[timestamp]] = paste(isoyear(result[[timestamp]]), sprintf("%02d", isoweek(result[[timestamp]])), sep = "-W")
     }
   }
   else if(timestampFormat == "month"){
-    result = result[result$lastunder <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," months"))[2], ]
+    result = result[result$lastunder <= as.Date(currentTime) %m-% months(thresholdTime), ]
     if(nrow(result)>0){
-      result$toohighinmonth = paste((year(result[[timestamp]]) - year(result$lastunder))*12 + (month(result[[timestamp]]) - month(result$lastunder)),"month")
-      result$lastunder = format(result$lastunder, "%Y-%m")
-      result[[timestamp]] = format(result[[timestamp]], "%Y-%m")
+      result$toohighinmonth = paste(interval(result$lastunder, result[[timestamp]]) %/% months(1), "months")
+      result$lastunder = paste(year(result$lastunder), sprintf("%02d", month(result$lastunder)), sep = "-")
+      result$mindate = paste(year(result$mindate), sprintf("%02d", month(result$mindate)), sep = "-")
+      result[[timestamp]] = paste(year(result[[timestamp]]), sprintf("%02d", month(result[[timestamp]])), sep = "-")
     }
   }
   else if(timestampFormat == "quarter"){
-    result = result[result$lastunder <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," quarters"))[2], ]
+    result = result[result$lastunder <= as.Date(currentTime) %m-% months(thresholdTime * 3), ]
     if(nrow(result)>0){
-      result$toohighinquarters = paste((year(result[[timestamp]]) - year(result$lastunder))*4 + (quarter(result[[timestamp]]) - quarter(result$lastunder)),"quaters")
+      result$toohighinquarters = paste(round((interval(result$lastunder, result[[timestamp]]) %/% months(1)) /3, digits = 2 ), "quarters")
       result$lastunder = paste0(year(result$lastunder), "-Q", quarter(result$lastunder))
+      result$mindate = paste0(year(result$mindate), "-Q", quarter(result$mindate))
       result[[timestamp]] = paste0(year(result[[timestamp]]), "-Q", quarter(result[[timestamp]]))
     }
   }
   else if(timestampFormat == "year"){
-    result = result[result$lastunder <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," years"))[2], ]
+    result = result[result$lastunder <= as.Date(currentTime) %m-% years(thresholdTime), ]
     if(nrow(result)>0){
       result$toohighinyears = paste(year(result[[timestamp]]) - year(result$lastunder),"years")
       result$lastunder = year(result$lastunder)
+      result$mindate = year(result$mindate)
       result[[timestamp]] = year(result[[timestamp]])
     }
   }
@@ -354,6 +327,7 @@ computeOverperformer = function(data,
 
     equals <- which(result$lastunder == result$mindate)
 
+    result <- data.frame(result)
     result$lastunder<- as.character(result$lastunder)
     result$lastunder[equals] <- paste(">=", result$lastunder[equals])
     result[, ncol(result)]<- as.character( result[, ncol(result)])
@@ -378,6 +352,7 @@ computeOverperformer = function(data,
 #' @param timestampFormat Declares in which format the timestamp comes in (i.e., "day", "week", "month", "quarter", "year")
 #' @param currentTime Qualifying date for the value variable. Date must exist in data and have the same format as timestamp-variable.
 #' @param thresholdTime Time for which the value shouldn't exceed the threshold value. Number declares the time in the format of timestampFormat.
+#' @param use_latest boolean value. If TRUE data will expand and dates with noexisting values will be filled up with the latest known values.
 #' @return Returns a data frame listing all constant items, the date since when the stock is constant and the value of the stock since this time.
 #' @seealso \code{\link{computeUnderperformer}} \code{\link{computeOverperformer}}
 #' @author Leon Binder \email{leon.binder@@th-deg.de}
@@ -391,7 +366,8 @@ computeOverperformer = function(data,
 #'                              timestamp = "date",
 #'                              timestampFormat = "day",
 #'                              currentTime = "2019-07-27",
-#'                              thresholdTime = 7)
+#'                              thresholdTime = 7,
+#'                              use_latest = FALSE)
 #' @export
 computeConstants = function(data,
                             value,
@@ -399,24 +375,24 @@ computeConstants = function(data,
                             timestamp,
                             timestampFormat = c("day", "week", "month", "quater","year"),
                             currentTime,
-                            thresholdTime = 7) {
-
+                            thresholdTime = 7,
+                            use_latest = FALSE) {
+  
+  if(use_latest){
+    data<-expandData(data = data,
+                     expand = group,
+                     expandTo = "all",
+                     valueColumns = value,
+                     latest_values = T,
+                     timestamp = timestamp,
+                     timestampFormat = timestampFormat,
+                     keepData = F)
+  }
 
   ### Bring timestamp and currentTime to date format
-  if (timestampFormat == "day") {
-    data[[timestamp]] = as.Date(as.character(data[[timestamp]]))
-  }
-  else if (timestampFormat == "week") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-1")
-    data[[timestamp]] = ISOweek2date(data[[timestamp]])
-    currentTime = paste0(currentTime, "-1")
-    currentTime<-ISOweek2date(currentTime)
-  }
-  else if (timestampFormat == "month") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01")
-    currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
+  if (timestampFormat %in% c("day", "week", "month", "year")) {
+    data[[timestamp]] = as.Date(parse_iso_8601(data[[timestamp]]))
+    currentTime = as.Date(parse_iso_8601(currentTime))
   }
   else if (timestampFormat == "quarter") {
     data[[timestamp]] = paste0(substr(data[[timestamp]],1,4), "-", as.numeric(substr(data[[timestamp]],7,7))*3, "-01")
@@ -424,52 +400,41 @@ computeConstants = function(data,
     currentTime = paste0(substr(currentTime,1,4), "-", as.numeric(substr(currentTime,7,7))*3, "-01")
     currentTime<-as.Date(currentTime, format = "%Y-%m-%d")
   }
-  else if (timestampFormat == "year") {
-    data[[timestamp]] = paste0(data[[timestamp]], "-01-01")
-    data[[timestamp]] = as.Date(data[[timestamp]], format = "%Y-%m-%d")
-    currentTime = paste0(currentTime, "-01-01")
-    currentTime = as.Date(currentTime, format = "%Y-%m-%d")
-  }
 
   ### Select only Data in the past of currentTime
   data = data[data[[timestamp]] <= currentTime, ]
   data = data[order(data[[timestamp]]),]
+  
   ### Determine last timestamp when thresholdValue was overtaken
 
   last_change<-setDT(data)[,list(no_change_since = get(timestamp)[length(get(timestamp))- last(rle(get(..value))$length) + 1],
                                  value = last(rle(get(..value))$value)),
                            by= eval(get("group"))]
 
-
-  ### Determine date since when the data at least should be constant
-
-  startdate = as.Date(currentTime) - thresholdTime
-  result = last_change[last_change$no_change_since <= startdate, ]
-
   ### Subset data for condition thresholdTime and determine timespan in aggregation of timestampFormat
   if(timestampFormat == "day"){
     result = last_change[last_change$no_change_since <= as.Date(currentTime) - thresholdTime, ]
   }
   else if(timestampFormat =="week"){
-    result = last_change[last_change$no_change_since <= as.Date(currentTime) - as.difftime(thresholdTime, units="weeks"), ]
+    result = last_change[last_change$no_change_since <= as.Date(currentTime) %m-% weeks(thresholdTime), ]
     if(nrow(result)>0){
-      result$no_change_since = ISOweek(result$no_change_since)
+      result$no_change_since = paste(isoyear(result$no_change_since), sprintf("%02d", isoweek(result$no_change_since)), sep = "-W")
     }
   }
   else if(timestampFormat == "month"){
-    result = last_change[last_change$no_change_since <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," months"))[2], ]
+    result = last_change[last_change$no_change_since <= as.Date(currentTime) %m-% months(thresholdTime), ]
     if(nrow(result)>0){
-      result$no_change_since = format(result$no_change_since, "%Y-%m")
+      result$no_change_since = paste(year(result$no_change_since), sprintf("%02d", month(result$no_change_since)), sep = "-")
     }
   }
   else if(timestampFormat == "quarter"){
-    result = last_change[last_change$no_change_since <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," quarters"))[2], ]
+    result = last_change[last_change$no_change_since <= as.Date(currentTime) %m-% months(thresholdTime*3), ]
     if(nrow(result)>0){
-      result$no_change_since = paste0(year(result$no_change_since), "-Q", quarter(result$no_change_since))
+      result$no_change_since = paste(year(result$no_change_since), quarter(result$no_change_since), sep = "-Q")
     }
   }
   else if(timestampFormat == "year"){
-    result = last_change[last_change$no_change_since <= seq(as.Date(currentTime), length = 2, by = paste0("-",thresholdTime," years"))[2], ]
+    result = last_change[last_change$no_change_since <= as.Date(currentTime) %m-% years(thresholdTime), ]
     if(nrow(result)>0){
       result$no_change_since = year(result$no_change_since)
     }